LayoutTests/http/tests/security/cross-origin-css-in-quirks-4.html

<html>
<!--
    Note the lack of doctype above: That is intentional. This test
    MUST be in quirks mode to have any real testing power.
-->
<head>
<title>Cross-origin CSS</title>

<script src="/w3c/resources/testharness.js"></script>
<script src="/w3c/resources/testharnessreport.js"></script>

<link rel="stylesheet" href="resources/xorigincss3.html"></link>

<script>
function getBackgroundColorForId(id) {
    return window.getComputedStyle(document.getElementById(id), null).getPropertyValue('background-color')
}

var onloadTest = async_test("Testing cross-origin and MIME behavior for CSS.");

window.onload = function () {
    test(function () {
        assert_equals(getBackgroundColorForId('id4'), 'rgb(255, 255, 0)');
    }, 'xorigincss3.html should be loaded via &lt;link&gt;');

    onloadTest.done();
};
</script>
</head>
<body>
    <div id="id4" class="id4"></div>
</body>
</html>
